+++ lophine-server/luminol-patches/features/0006-Carpet-features.patch
@@ -5,7 +5,7 @@ Subject: [PATCH] Carpet features
 
 
 diff --git a/src/main/java/me/earthme/luminol/config/ConfigManager.java b/src/main/java/me/earthme/luminol/config/ConfigManager.java
-index f5fb4e210c07f32e52eb8c1ba166d1571a6778c7..cb8f143b5530e30b96d5af3b23374ee13b5a4497 100644
+index f5fb4e210c07f32e52eb8c1ba166d1571a6778c7..e990c147e1b632ef5a33e4f3c37b5cd608040ac7 100644
 --- a/src/main/java/me/earthme/luminol/config/ConfigManager.java
 +++ b/src/main/java/me/earthme/luminol/config/ConfigManager.java
 @@ -25,6 +25,13 @@ public class ConfigManager {
@@ -17,7 +17,7 @@ index f5fb4e210c07f32e52eb8c1ba166d1571a6778c7..cb8f143b5530e30b96d5af3b23374ee1
 +                "lophine_carpet",
 +                "lophine_carpet_config.toml",
 +                "lophinecarpetconfig",
-+                "fun.bm.lophine.config.carpet.modules"
++                "fun.bm.lophine.carpet.config.modules"
 +        )); // add lophine carpet config
          preLoad();
      }
