+++ lophine-server/src/main/java/fun/bm/lophine/carpet/config/modules/CoreConfig.java
@@ -1,6 +1,6 @@
-package fun.bm.lophine.config.carpet.modules;
+package fun.bm.lophine.carpet.config.modules;
 
-import fun.bm.lophine.config.carpet.CarpetCompatSync;
+import fun.bm.lophine.carpet.CarpetCompatSync;
 import me.earthme.luminol.config.IConfigModule;
 import me.earthme.luminol.config.flags.ConfigClassInfo;
 import me.earthme.luminol.config.flags.ConfigInfo;
