+++ lophine-server/src/main/java/fun/bm/lophine/carpet/config/modules/GeneralCompatConfig.java
@@ -1,4 +1,4 @@
-package fun.bm.lophine.config.carpet.modules;
+package fun.bm.lophine.carpet.config.modules;
 
 import me.earthme.luminol.config.IConfigModule;
 import me.earthme.luminol.config.flags.ConfigClassInfo;
