+++ lophine-server/src/main/java/org/leavesmc/leaves/bot/agent/actions/ServerUseItemToAction.java
@@ -17,7 +17,7 @@
 
 package org.leavesmc.leaves.bot.agent.actions;
 
-import fun.bm.lophine.config.carpet.modules.FakePlayerCompatConfig;
+import fun.bm.lophine.carpet.config.modules.FakePlayerCompatConfig;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.Entity;
